refactor(core/command): 重命名 `parse-command.ts`
`command-parser.ts` → `parse-command.ts`

动宾结构符合函数式模块命名规范
