+++ src/core/command/index.ts
@@ -1,5 +1,5 @@
 export * from './command';
 export * from './manager';
-export * from './command-parser';
+export * from './parse-command';
 export * from './errors';
 export type * from './types';
