+++ src/core/command/parse-command.ts
